added new functionality
if whipped cream is selected with any cone ice cream, it gives the following error: "You cannot have whippped cream on a cone"
removed checkTopping function and replaced it with checkWhip function. Created getSelectedRadio(radioArray) function that create array of radio input elment with querySelectorAll and return that checked button. If whip checked and all but cup is selected then it adds the above error.