Added an end-point (i.e., a route) for a POST , post the coupon code to '/coupon-check' using Ajax.
In App.js file, Added an end-point (i.e., a route) for a POST request to '/coupon-check' that responds with one of the following:
'<span class="valid">Coupon validated</span>'
'<span class="invalid">Invalid coupon</span>'

in Ajax.js file. The parameter is coupon and the value is couponCode (as defined in the beginning of the function).Wrote the code to post the coupon code to '/coupon-check' using Ajax.